add fix for schedule
