+++ Tracker/Models/Tracker.swift
@@ -7,14 +7,24 @@
 
 import Foundation
 
+enum WeekDay: String, CaseIterable, Codable {
+    case monday = "Понедельник"
+    case tuesday = "Вторник"
+    case wednesday = "Среда"
+    case thursday = "Четверг"
+    case friday = "Пятница"
+    case saturday = "Суббота"
+    case sunday = "Воскресенье"
+}
+
 final class Tracker: Codable {
     let id: Int
     let name: String
     let color: String
     let emodji: String
-    let timetable: String
+    let timetable: [WeekDay]
 
-    init(id: Int, name: String, color: String, emodji: String, timetable: String) {
+    init(id: Int, name: String, color: String, emodji: String, timetable: [WeekDay]) {
         self.id = id
         self.name = name
         self.color = color
